external workload: fold GCV2 abort utility
