+++ pkg/extworkload/util.go
@@ -15,8 +15,14 @@
 package extworkload
 
 import (
+	"context"
+	"os"
+
 	"github.com/pingcap/tidb/pkg/config"
 	"github.com/pingcap/tidb/pkg/kv"
+	"github.com/pingcap/tidb/pkg/util/intest"
+	"github.com/pingcap/tidb/pkg/util/logutil"
+	"go.uber.org/zap"
 )
 
 // IsEnabled reports whether a Manager is present.
@@ -55,6 +61,23 @@ func GetManagerFromStore(store kv.Storage) Manager {
 	}
 	mgr, _ := v.(Manager)
 	return mgr
+}
+
+// AbortGCV2ForUpgrade aborts GCV2 work and exits when this TiDB is the dedicated
+// GCV2 worker, which must not run normal bootstrap upgrade work.
+func AbortGCV2ForUpgrade(store kv.Storage) {
+	mgr := GetManagerFromStore(store)
+	if !IsGCV2Worker(mgr) {
+		return
+	}
+	if err := mgr.AbortGCV2(context.Background()); err != nil {
+		logutil.BgLogger().Fatal("abort GCV2 worker failed", zap.Error(err))
+	}
+	logutil.BgLogger().Info("GCV2 worker aborted")
+	if intest.InTest {
+		return
+	}
+	os.Exit(0)
 }
 
 func roleIs(m Manager, role config.ExternalWorkloadRole) bool {
